service: move DaemonController from nydus into nydus-service
Move DaemonController from nydus into nydus-service so it can
be reused by other project.
