+++ src/bin/nydusd/api_server_glue.rs
@@ -384,7 +384,7 @@ impl ApiServerController {
         let api_server = ApiServer::new(to_router)?;
         let api_handler = ApiServerHandler::new(api_server, from_router)?;
         let (router_thread, waker) = start_http_thread(apisock, None, to_handler, from_handler)?;
-        let daemon_waker = DAEMON_CONTROLLER.waker.clone();
+        let daemon_waker = DAEMON_CONTROLLER.alloc_waker();
 
         info!("HTTP API server running at {}", apisock);
         let handler_thread = std::thread::Builder::new()
